+++ src/ladb_opencutlist/js/plugins/jquery.ladb.dialog.js
@@ -157,8 +157,6 @@
 
     LadbDialog.prototype.selectTab = function (tabName, callback) {
 
-        console.log(tabName, callback);
-
         var $tab = this.tabs[tabName];
         var $freshTab = false;
         if (tabName != this.activeTabName) {
@@ -229,8 +227,6 @@
     };
 
     LadbDialog.prototype.executeCommandOnTab = function(tabName, command, parameters, callback) {
-
-        console.log(tabName, command, parameters, callback);
 
         // Select tab and execute command
         this.selectTab(tabName, function($tab) {