+++ src/ladb_opencutlist/ruby/constants.rb
@@ -1,7 +1,7 @@
 module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
-    EXTENSION_VERSION = '1.6.0'.freeze
-    EXTENSION_BUILD = '201902111715'.freeze
+    EXTENSION_VERSION = '1.6.1'.freeze
+    EXTENSION_BUILD = '201902151208'.freeze
 
 end